% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bio_Cust_Rec_Display.R
\name{bio_Cust_Rec_Display}
\alias{bio_Cust_Rec_Display}
\title{Custom Display of NSSP ED Patient Record}
\usage{
bio_Cust_Rec_Display(df, UID, colUID = NULL, colList = NULL)
}
\arguments{
\item{df}{a data.frame of patient records}

\item{UID}{the index number of the row to be displayed}

\item{colUID}{the index number of the column that contains the unique row
identifiers}

\item{colList}{a vector of names of columns to be displayed.}
}
\value{
a list containing information for a single record
}
\description{
Custom Display of NSSP ED Patient Record
}
\details{
This function displays a selected record as a list. It needs a
  data.frame and a unique identifier for the record. The function can also
  display a record based on other unique identifiers such as the row record
  number. In that case the index number of the column with the unique
  identifier using the colUID argument. The number of columns displayed may
  also be reduced by providing a vector with the name of the columns of
  interest using the colList argument.
  This function is particularly useful when inspecting and labeling cases.
}

