% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reviewProcs.R
\name{createReviewSet}
\alias{createReviewSet}
\title{Creates a data.frame containing records for manual review and labeling}
\usage{
createReviewSet(df, labcol = NULL, colList = NULL, savename = NULL)
}
\arguments{
\item{df}{a data.frame containing the raw data to be prepared for review}

\item{labcol}{a string indicating the name of the label column}

\item{colList}{a vector of column names to be included in the final review
data.frame.}

\item{savename}{a string (without extension) indicating the name of the
review data.frame to be saved on the hard drive.}
}
\value{
a data.frame with a label column at the end.
}
\description{
Creates a data.frame containing records for manual review and labeling
}
\details{
If the savename is not provided the file will be saved in the
  working directory. The savename can contain the full path of the file.
}

