% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterBioByWords.R
\name{filterBioByWords}
\alias{filterBioByWords}
\title{Selects records containing the inclusion keywords if they do not contain the
exclusion keywords}
\usage{
filterBioByWords(df, inclusions = NULL, exclusions = NULL)
}
\arguments{
\item{df}{a processed NSSP data.frame}

\item{inclusions}{a string of keywords indicative of the condition of
interest}

\item{exclusions}{a string of keywords if found in the records with inclusion
keywords disqualify those previously selected records}
}
\value{
a list containing 3 elements. cut1, a data.frame containing records
  meeting the inclusion criteria. cut2, a data.frame containing the records
  meeting the exclusion criteria within cut1. cut3, a data.frame containing
  the final selection in the process. cut3 is a data.frame resulting from the
  subtraction of cut2 from cut1.
}
\description{
Selects records containing the inclusion keywords if they do not contain the
exclusion keywords
}

