% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.NSSP.data.R
\name{fixDateTime}
\alias{fixDateTime}
\title{Converts Date Strings into R Date/Time Format}
\usage{
fixDateTime(datdate, dateType = "date")
}
\arguments{
\item{datdate}{a string representing a date and/or time. NOTE: This
implementation is incomplete.}

\item{dateType}{a string indicating the type of the date/time represented by
the string to be converted.}
}
\value{
an R datetime object
}
\description{
Converts Date Strings into R Date/Time Format
}
\details{
This implementation is inclomplete at this time. For datdate
  variable the acceptable values are 'date' and 'datetime'. Other options are
  forthcomming.
}

