% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flagBioRecs.R
\name{flagBioRecs}
\alias{flagBioRecs}
\title{Flags records based on set keywords and specified columns in a data.frame}
\usage{
flagBioRecs(df, terms, searchfields = NULL)
}
\arguments{
\item{df}{a NSSP data.frame with original column names}

\item{terms}{a vector of stings and words of interest}

\item{searchfields}{the index of the columns targeted for the search}
}
\value{
a table with the Unique_Visiting_ID as the first column and a column
  for each field where one or more of the terms have been found.
}
\description{
Flags records based on set keywords and specified columns in a data.frame
}
\details{
In the final table, each field will receive a 1 if there is a hit or
  a 0 otherwise. If a target column doesn't contain any of the sought terms
  it does not get included in the final table. It helps identify which field
  is the most relevant for this particular terms string.
}

