% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flagBioRecs.R
\name{getDefinitions}
\alias{getDefinitions}
\title{Retrieves a specific inclusion terms or exclusion terms for case definitions}
\usage{
getDefinitions(condition, incl_excl, default_utilities = TRUE, date = NULL)
}
\arguments{
\item{condition}{a string indicating the name of the condition of interest}

\item{incl_excl}{a string indicating "inclusions" or "exclusions"}

\item{default_utilities}{a boolean variable indicating if the default file
containing the inclusion/exclusion terms should be used or not.}

\item{date}{a single R date indicating the signature date of the definition}
}
\value{
A vector of one string that represents the search term
}
\description{
Retrieves a specific inclusion terms or exclusion terms for case definitions
}
\details{
The date part of this function is not fully implemented yet. It is
  important that the names of the condition is spelled correctly.
}

