% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryofevents.R
\name{summaryofEvents}
\alias{summaryofEvents}
\title{Summarizes the visits by age groups and by day}
\usage{
summaryofEvents(condition, df, fields)
}
\arguments{
\item{condition}{a string indicating the condition under consideration}

\item{df}{a data.frame containing individual records with date and age
variables formatted properly.}

\item{fields}{a vector of column indexes to be included in the final output}
}
\value{
a list containg the raw reduced data.frame, a data.frame containing
  the number of events for each day, and a third data.frame containing the
  number of events per facility
}
\description{
Summarizes the visits by age groups and by day
}
\details{
The raw data.frame can be accessed using the label "tmpdf". The
  events per day data.frame can be accessed with the label "events.per.day".
  The number of events per day for a facility can be accessed with the label
  "events.per.day.per.facility".

  Note: fields parameter must be an integer vector. Otherwise it will be
  ignored. In addition, the Unique visit ID field must also be included,
  otherwise an error will occur.

  Note: Other columns that are required in the fields vector includes Earliest_Date, Age,
}

