% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.NSSP.data.R
\name{tidy_BioSense}
\alias{tidy_BioSense}
\title{Tidies-up Raw NSSP Data}
\usage{
tidy_BioSense(df)
}
\arguments{
\item{df}{a raw NSSP data.frame}
}
\value{
Returns a data frame with additional calculated fields.
}
\description{
Tidies-up Raw NSSP Data
}
\details{
This function converts date and time fields to the right format for R. Also, it adds new calculated fields including one for the age of the patient in years (Age_years) and another one containing the concatenation of the literal fields (Text). In addition, it attempts to standardize the entries in the Age_Units field.
}

